+++ packages/pwa-kit-create-app/scripts/create-mobify-app.js
@@ -167,11 +167,8 @@ const runGenerator = (answers, {outputDir}) => {
         ]
     })
 
-    const PWAKitConfigJsonTemplate = require(`../assets/pwa/pwa-kit.config`)
-    const PWAKitConfigJsonPath = p.resolve(outputDir, 'pwa-kit.config.json')
-    writeJson(PWAKitConfigJsonPath, PWAKitConfigJsonTemplate)
+    const PWAKitConfigJsonTemplate = require(`../assets/pwa/default`).template
 
-    const APIConfigTemplate = require(`../assets/pwa/api.config`).template
     const commerceApi = {
         proxyPath: answers['scaffold-pwa'].mobify.ssrParameters.proxyConfigs[0].path,
         clientId: answers['commerce-api'].clientId,
@@ -185,8 +182,8 @@ const runGenerator = (answers, {outputDir}) => {
         siteId: answers['einstein-api'].siteId || answers['commerce-api'].siteId
     }
 
-    new sh.ShellString(APIConfigTemplate({commerceApi, einsteinApi})).to(
-        p.resolve(outputDir, 'app', 'api.config.js')
+    new sh.ShellString(PWAKitConfigJsonTemplate({commerceApi, einsteinApi})).to(
+        p.resolve(outputDir, 'config', 'default.js')
     )
 
     npmInstall(outputDir)
@@ -289,25 +286,7 @@ const buildAnswers = ({
         globals: {projectId},
         'scaffold-pwa': {
             name: projectId,
-            version: GENERATED_PROJECT_VERSION,
-            mobify: {
-                ssrParameters: {
-                    proxyConfigs: [
-                        {
-                            path: 'api',
-                            host: `${shortCode}.api.commercecloud.salesforce.com`
-                        },
-                        {
-                            path: 'ocapi',
-                            host: new URL(instanceUrl).hostname
-                        },
-                        {
-                            path: 'einstein',
-                            host: 'api.cquotient.com'
-                        }
-                    ]
-                }
-            }
+            version: GENERATED_PROJECT_VERSION
         },
 
         'commerce-api': {clientId, siteId, organizationId, shortCode},