+++ packages/pwa/app/ssr.js
@@ -9,8 +9,8 @@
 const path = require('path')
 const {getRuntime, serveStaticFile} = require('pwa-kit-runtime/ssr/server/express')
 const {isRemote} = require('pwa-kit-runtime/utils/ssr-server')
+const {getConfig} = require('pwa-kit-runtime/utils/ssr-config')
 const helmet = require('helmet')
-const pkg = require('../package.json')
 
 const options = {
     // The build directory (an absolute path)
@@ -27,7 +27,7 @@ const options = {
     manifestPath: 'static/manifest.json',
 
     // This is the value of the 'mobify' object from package.json
-    mobify: pkg.mobify,
+    mobify: getConfig(),
 
     // The port that the local dev server listens on
     port: 3000,