+++ packages/pwa-kit-runtime/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v2.0.0-dev.2 (Feb 10, 2022)
 
+- Add environment specific configuration support via `getConfig`. [#447](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/447)
 - Remove legacy remote proxy, which allowed remote environments to use proxy configs in package.json [#425](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/425)
 - Remove default `body-parser` middleware from express server.